+++ NEWS.md
@@ -1,3 +1,10 @@
+# broomstick 0.1.2.9100 (2019-06-12)
+
+## Minor Update
+
+* draft `augment` for `rpart`.
+* use markdown roxygen
+
 # broomstick 0.1.2.9000 (2019-01-30)
 
 ## Minor Update